Rename datasets/dataset.py to datasets/spectrum.py
