+++ gammapy/datasets/spectrum.py
