+++ gammapy/spectrum/reflected.py
@@ -348,7 +348,7 @@ class ReflectedRegionsBackgroundMaker:
         dataset_on_off : `SpectrumDatasetOnOff`
             On off dataset.
         """
-        from gammapy.datasets.dataset import SpectrumDatasetOnOff
+        from gammapy.datasets import SpectrumDatasetOnOff
 
         counts_off, acceptance_off = self.make_counts_off(dataset, observation)
